:tada: Add initial gamepad input class
